od: Solved warning messages for ArrayMath namespace

+++ src/Attributes/frequencyattrib.cc
@@ -238,7 +238,7 @@ bool Frequency::computeData( const DataHolder& output, const BinID& relpos,
 	}
 
 	if ( window_ ) window_->apply( myself->signal_ );
-	ArrayMath::removeBias<float_complex,float_complex,double>(
+	ArrayMath::removeBias<float_complex,float_complex,float>(
 							  *myself->signal_ );
 	for ( int idy=0; idy<sz; idy++ )
 	    myself->timedomain_->set( sz+idy, signal_->get(idy) );
